list Detail app

+++ src/main/java/com/example/Project/List/ListRepository.java
@@ -7,7 +7,8 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import java.util.Optional;
 
-public interface ListRepository extends JpaRepository<ListMain, Long> {
+public interface ListRepository extends JpaRepository<ListMain, Integer> {
     @Query("SELECT e FROM ListMain e ORDER BY e.createDate DESC")
     List<ListMain> findAllOrderByCreateDateDesc();
+
 }
